# functions/api/schema/openapi.yaml
openapi: '3.1.0'
info:
  title: OpenPecha API v2
  version: '2.1.0'
servers:
  - url: https://api-l25bgmwqoa-uc.a.run.app
    description: Development
  - url: https://api-aq25662yyq-uc.a.run.app
    description: Production
paths:
  /v2/instances/{instance_id}:
    get:
      summary: Retrieve instance by instance ID
      description: Fetch instance content for a given instance ID with optional alignment
      tags:
        - Instances
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the instance to retrieve
        - name: aligned
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether to include aligned source text
      responses:
        "200":
          description: Instance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  target:
                    type: object
                    properties:
                      pecha:
                        type: object
                      annotations:
                        type: array
                        items:
                          type: object
                  source:
                    type: object
                    properties:
                      pecha:
                        type: object
                      annotations:
                        type: array
                        items:
                          type: object
              examples:
                success:
                  summary: Text with alignment
                  value:
                    target:
                      pecha:
                        id: "I12345678"
                        base_text: "This is the target text content"
                      annotations:
                        - id: "A12345678"
                          type: "segmentation"
                    source:
                      pecha:
                        id: "I87654321"
                        base_text: "This is the source text content"
                      annotations:
                        - id: "A87654321"
                          type: "alignment"
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /v2/instances/{instance_id}/translation:
    post:
      summary: Create a translation
      description: Create a translation of an existing instance
      tags:
        - Instances
      parameters:
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the original instance to translate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - language
                - content
                - title
                - translator
                - translation_annotation
              properties:
                language:
                  type: string
                  description: Target language code
                content:
                  type: string
                  description: The translated text content
                title:
                  type: string
                  description: Title of the translation
                alt_titles:
                  type: array
                  items:
                    type: string
                translator:
                  type: object
                  properties:
                    person_id:
                      type: string
                    person_bdrc_id:
                      type: string
                    ai_id:
                      type: string
                translation_annotation:
                  type: array
                  items:
                    type: object
                  description: Alignment annotation for translation
                original_annotation:
                  type: array
                  items:
                    type: object
                  description: Optional alignment annotation for original text
                copyright:
                  type: string
                  enum: [public, copyrighted]
                  default: public
            examples:
              valid_translation:
                summary: Valid translation creation request
                value:
                  language: "en"
                  content: "This is the translated text content"
                  title: "Translated Title"
                  translator:
                    person_id: "P12345678"
                  translation_annotation:
                    - span:
                        start: 0
                        end: 20
                      index: 0
                      alignment_index: [0]
                  copyright: "public"
      responses:
        "201":
          description: Translation created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - "Translation created successfully"
                  id:
                    type: string
                    examples:
                      - "ABC12345678"
                  text_id:
                    type: string
                    examples:
                      - "ABC12345678"
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "422":
          $ref: '#/components/responses/ValidationError'
        "500":
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    InvalidRequest:
      description: There was an error with the request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
                examples:
                  - "There was an error with the request"
            required:
              - error
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
                examples:
                  - "Resource was not found"
            required:
              - error
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
            required:
              - error
              - details
          examples:
            fieldMissing:
              value:
                error: "Validation error"
                details: [
                  {
                    "loc": ["body", "title"],
                    "msg": "field required", 
                    "type": "value_error.missing"
                  }
                ]
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
                examples:
                  - "There was an error on the server"
            required:
              - error